+++ test/unit/mongo_db_admin/main.py
@@ -28,6 +28,167 @@ import version
 __version__ = version.__version__
 
 
+class ArgParser(object):
+
+    """Class:  ArgParser
+
+    Description:  Class stub holder for gen_class.ArgParser class.
+
+    Methods:
+        __init__
+        arg_cond_req
+        arg_file_chk
+        arg_valid_val
+        arg_cond_req_or
+        arg_dir_chk
+        arg_require
+        get_val
+        arg_xor_dict
+
+    """
+
+    def __init__(self):
+
+        """Method:  __init__
+
+        Description:  Class initialization.
+
+        Arguments:
+
+        """
+
+        self.cmdline = None
+        self.args_array = dict()
+        self.opt_req = None
+        self.opt_req2 = True
+        self.dir_perms_chk = None
+        self.dir_perms_chk2 = True
+        self.dir_chk = None
+        self.opt_con_or = None
+        self.opt_con_or2 = True
+        self.opt_xor_val = None
+        self.opt_xor_val2 = True
+        self.file_perm_chk = None
+        self.file_crt = None
+        self.arg_file_chk2 = True
+        self.opt_valid_val = None
+        self.arg_valid_val2 = True
+        self.opt_con_req = None
+        self.opt_con_req2 = True
+
+    def arg_cond_req(self, opt_con_req):
+
+        """Method:  arg_cond_req
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_cond_req.
+
+        Arguments:
+
+        """
+
+        self.opt_con_req = opt_con_req
+
+        return self.opt_con_req2
+
+    def arg_file_chk(self, file_perm_chk, file_crt):
+
+        """Method:  arg_file_chk
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_file_chk.
+
+        Arguments:
+
+        """
+
+        self.file_perm_chk = file_perm_chk
+        self.file_crt = file_crt
+
+        return self.arg_file_chk2
+
+    def arg_valid_val(self, opt_valid_val):
+
+        """Method:  arg_valid_val
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_valid_val.
+
+        Arguments:
+
+        """
+
+        self.opt_valid_val = opt_valid_val
+
+        return self.arg_valid_val2
+
+    def arg_cond_req_or(self, opt_con_or):
+
+        """Method:  arg_cond_req_or
+
+        Description:  Method stub holder for
+            gen_class.ArgParser.arg_cond_req_or.
+
+        Arguments:
+
+        """
+
+        self.opt_con_or = opt_con_or
+
+        return self.opt_con_or2
+
+    def arg_dir_chk(self, dir_perms_chk):
+
+        """Method:  arg_dir_chk
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_dir_chk.
+
+        Arguments:
+
+        """
+
+        self.dir_perms_chk = dir_perms_chk
+
+        return self.dir_perms_chk2
+
+    def arg_require(self, opt_req):
+
+        """Method:  arg_require
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_require.
+
+        Arguments:
+
+        """
+
+        self.opt_req = opt_req
+
+        return self.opt_req2
+
+    def get_val(self, skey, def_val):
+
+        """Method:  get_val
+
+        Description:  Method stub holder for gen_class.ArgParser.get_val.
+
+        Arguments:
+
+        """
+
+        return self.args_array.get(skey, def_val)
+
+    def arg_xor_dict(self, opt_xor_val):
+
+        """Method:  arg_xor_dict
+
+        Description:  Method stub holder for gen_class.ArgParser.arg_xor_dict.
+
+        Arguments:
+
+        """
+
+        self.opt_xor_val = opt_xor_val
+
+        return self.opt_xor_val2
+
+
 class ProgramLock(object):
 
     """Class:  ProgramLock
@@ -94,8 +255,11 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        self.args_array = {"-c": "CfgFile", "-d": "CfgDir"}
-        self.args_array2 = {"-c": "CfgFile", "-d": "CfgDir", "-y": "Flavor"}
+        self.args = ArgParser()
+        self.args2 = ArgParser()
+        self.args.args_array = {"-c": "CfgFile", "-d": "CfgDir"}
+        self.args2.args_array2 = {
+            "-c": "CfgFile", "-d": "CfgDir", "-y": "Flavor"}
         self.proglock = ProgramLock(["cmdline"], "FlavorID")
 
     @mock.patch("mongo_db_admin.gen_libs.help_func")