+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Upgraded python-lib to v2.10.1
 - Replaced the arg_parser code with gen_class.ArgParser code.
 
+### Changed
+- Multiple functions: Replaced the arg_parser code with gen_class.ArgParser code.
+- main: Removed gen_libs.get_inst call.
+
 
 ## [2.4.4] - 2023-06-07
 - Updated mongo-lib to v4.2.4